+++ inet/protocols/rwhod.h
@@ -36,13 +36,15 @@
 #ifndef _RWHOD_H_
 #define	_RWHOD_H_
 
+#include <sys/types.h>
+
 /*
  * rwho protocol packet format.
  */
 struct	outmp {
 	char	out_line[8];		/* tty name */
 	char	out_name[8];		/* user id */
-	long	out_time;		/* time on */
+	int32_t	out_time;		/* time on */
 };
 
 struct	whod {
